Corrección del delete en CuponMontoFijo y agregado CuponPorcentaje

+++ SWCE.Web/Models/CuponMontoFijo/CreateCuponMontoFijoModel.cs
@@ -2,7 +2,7 @@
 {
     public class CreateCuponMontoFijoModel
     {
-        public int monto { get; set; }
+        public decimal monto { get; set; }
         public int id { get; set; }
         public DateTime fechaExpiracion { get; set; }
     }
